# api/openapi.json
{
  "components": {
    "schemas": {
...
      "FeedOut": {
        "description": "One upstream capture, as the Feed Health board reads it.\n\nA feed here IS a scheduled Redash query that captures into the historical\nwarehouse, so the id is the table it captures into and it matches the\ndataset of the same id one to one. That is structural rather than a\nconvention worth breaking: `get_or_create_table_name` keys the registry on\nquery_id and hands back exactly one table, so `dataset_count` cannot be\nanything but 1 while that holds.",
        "properties": {
          "alertId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alertid"
          },
          "cadence": {
            "title": "Cadence",
            "type": "string"
          },
          "cadenceSource": {
            "enum": [
              "declared",
              "schedule",
              "none"
            ],
            "title": "Cadencesource",
            "type": "string"
          },
          "datasetCount": {
            "title": "Datasetcount",
            "type": "integer"
          },
          "expectedIntervalSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedintervalseconds"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastReceivedAt": {
            "title": "Lastreceivedat",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "fresh",
              "stale"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "source",
          "cadence",
          "cadenceSource",
          "expectedIntervalSeconds",
          "alertId",
          "lastReceivedAt",
          "status",
          "datasetCount"
        ],
        "title": "FeedOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TagCountOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "count"
        ],
        "title": "TagCountOut",
        "type": "object"
      },
      "TagsIn": {
        "description": "The whole set for one object: a replace, not an add/remove pair.",
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "tags"
        ],
        "title": "TagsIn",
        "type": "object"
      },
      "TagsOut": {
        "description": "What is actually stored, after normalization: `  Rail  ` comes back `rail`.",
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "tags"
        ],
        "title": "TagsOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "veodyn-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/feeds": {
      "get": {
        "description": "Every scheduled capture the warehouse holds, with its cadence and source.\n\nThe two Redash lookups run as the CALLER, not as the service account, so a\ndata source or a query the reader cannot see goes unnamed rather than named\nfor them. Both are best-effort by construction: they label a list the\nwarehouse already produced, and neither can empty it.",
        "operationId": "list_feeds_feeds_get",
        "parameters": [
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedOut"
                  },
                  "title": "Response List Feeds Feeds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feeds",
        "tags": [
          "feeds"
        ]
      }
    },
...
    "/tags": {
      "get": {
        "description": "The org's tag vocabulary, unioned across every kind this service owns.\n\nThe same item shape Redash's QueryTagsResource returns, so the frontend\nmerges this with the query and dashboard vocabularies through one mapper.\nReserved `domain:` tags are excluded, so they can never be suggested.",
        "operationId": "list_tags_tags_get",
        "parameters": [
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagCountOut"
                  },
                  "title": "Response List Tags Tags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tags",
        "tags": [
          "tags"
        ]
      }
    },
    "/tags/{object_type}/{object_id}": {
      "put": {
        "description": "Replace the whole tag set for one object.\n\nReplace rather than add/remove because the frontend's TagsControl hands back\nthe full array on every change, which is idempotent under retry. Tags are\nnormalized here as well as in the browser, since this endpoint is reachable\nwithout the UI and matching is exact, so one unnormalized write would split a\ntag in two for everyone.",
        "operationId": "set_tags_tags__object_type___object_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "object_type",
            "required": true,
            "schema": {
              "title": "Object Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Tags",
        "tags": [
          "tags"
        ]
      }
    }
  }
}